+++ ledger/tests/test_api/test_char_ledger_api.py
@@ -52,39 +52,27 @@ class ManageApiLedgerCharEndpointsTest(TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json(), expected_data)
 
-    @patch("ledger.api.character.ledger.get_cache_stale")
-    @patch("ledger.api.character.ledger.IS_TESTING", False)
-    def test_get_character_ledger_api_cache_no_testing(self, mock_get_cache_stale):
+    def test_get_character_ledger_api_cache_no_testing(self):
         self.client.force_login(self.user)
         url = "/ledger/api/account/0/ledger/year/2024/month/3/"
-
-        mock_get_cache_stale.return_value = CharmonthlyMarch
 
         response = self.client.get(url)
         expected_data = CharmonthlyMarch
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json(), expected_data)
 
-    @patch("ledger.api.character.ledger.get_cache_stale")
-    @patch("ledger.api.character.ledger.IS_TESTING", True)
-    def test_get_character_ledger_api_no_cache_testing(self, mock_get_cache_stale):
+    def test_get_character_ledger_api_no_cache_testing(self):
         self.client.force_login(self.user)
         url = "/ledger/api/account/0/ledger/year/2024/month/3/"
-
-        mock_get_cache_stale.return_value = False
 
         response = self.client.get(url)
         expected_data = CharmonthlyMarch
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json(), expected_data)
 
-    @patch("ledger.api.character.ledger.get_cache_stale")
-    @patch("ledger.api.character.ledger.IS_TESTING", False)
-    def test_get_character_ledger_api_no_cache_no_testing(self, mock_get_cache_stale):
+    def test_get_character_ledger_api_no_cache_no_testing(self):
         self.client.force_login(self.user)
         url = "/ledger/api/account/0/ledger/year/2024/month/3/"
-
-        mock_get_cache_stale.return_value = False
 
         response = self.client.get(url)
         expected_data = CharmonthlyMarch